Add files via upload
The code you provided is an implementation of the backpropagation algorithm for a simple neural network. The neural network has three layers: an input layer with two nodes, a hidden layer with two nodes, and an output layer with one node. The code first initializes the weights and biases of the neural network, and then it enters a loop that repeats five times.

In each iteration of the loop, the code first calculates the values of the hidden layer nodes (z1 and z2). It then calculates the value of the output layer node (y). The code then calculates the error between the desired output (1) and the actual output (y). This error is used to calculate the gradients of the weights and biases. The gradients are then used to update the weights and biases.

After the loop has finished, the code prints the values of the weights, biases, and nodes.

Here is a more detailed explanation of the code:

The first line of code defines the variables x1, x2, e, learnrate, t, w11, w12, w21, w22, b1, b2, and b3. These variables represent the input values, the learning rate, the desired output, and the weights and biases of the neural network.
The second line of code prints the header of the table that will be used to store the results of the algorithm.
The third line of code starts the loop.
The fourth line of code calculates the values of the hidden layer nodes (z1 and z2).
The fifth line of code calculates the value of the output layer node (y).
The sixth line of code calculates the error between the desired output (1) and the actual output (y).
The seventh line of code calculates the gradients of the weights and biases.
The eighth line of code updates the weights and biases.
The ninth line of code prints the values of the weights, biases, and nodes.
The tenth line of code increments the counter i.
The eleventh line of code checks if i is less than or equal to 5. If it is, the loop repeats. If it is not, the loop terminates.
The backpropagation algorithm is a powerful tool for training neural networks. It allows the neural network to learn from its mistakes and to improve its performance over time. The code you provided is a simple implementation of the backpropagation algorithm, but it can be used to train neural networks to solve a variety of problems.